feat(backup): split settings page into tabs (refs #108)
The GitHub backup settings page stacked connection, target/schedule and
activity cards into one long scroll. Split it into four tabs reusing the
existing api-tab look: Connection, Target & Schedule, Activity, Sync Log.

- Extract the Target & Schedule form from BackupSection into
  BackupTargetCard (owns form state; unsaved edits survive tab switches)
- Extract the sync log from BackupActivityCard into BackupLogCard with
  its own refresh and empty state
- Keep all tab panels mounted (hidden, not unmounted) so a pending
  device-flow login keeps polling across tab switches
- Land on the first tab that still needs attention (connect -> target ->
  activity) and mark the Activity tab with a red dot when unhealthy
- Refresh the sync log after a manual "Back up all now" run
